requirements.txt
# True Learning System - Python依赖

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 模板和表单
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.0

# AI API
openai>=1.0.0

# 数据库
sqlalchemy>=2.0.23

# 数据验证
pydantic>=2.5.0

# 环境变量
python-dotenv>=1.0.0

# 文档解析
pypdf>=4.2.0

# Testing
pytest>=9.0.0
pytest-asyncio>=0.23.0
